requirements.txt
fastapi>=0.110.0
uvicorn>=0.21.0
fastembed==0.6.1

elasticsearch==8.17.2
python-dotenv>=1.0.0
pydantic>=1.10.7
aiofiles>=23.1.0
python-multipart>=0.0.6
httpx>=0.24.0
numpy>=1.24.3
aiohttp>=3.8.0

# Google Cloud Vertex AI
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
pillow>=10.0.0

# mcp
mcp-use==1.2.8

# open ai
openai==1.77.0
langchain-openai==0.3.16

# jwt
jwt==1.3.1
PyJWT==2.10.1